admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Nico Romero <ika.22.94@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
Flask-Session = "^0.4.0"
Flask-Bcrypt = "^1.0.1"
PyJWT = "^2.6.0"
Flask-JWT-Extended = "^4.4.4"
dateutils = "^0.6.12"
Flask-Cors = "^3.0.10"
qrcode = "^7.3.1"
pillow = "^9.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
flask-shell-ipython = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
